poko/requirements.txt
anyio==4.2.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.0.5
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
black==24.2.0
bleach==6.1.0
certifi==2024.2.2
# cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cycler==0.11.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.6
django-allauth==0.61.1
djangorestframework-simplejwt==5.3.1
dj-rest-auth==6.0.0
django-cors-headers==4.4.0
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-pandas==0.6.6
djangorestframework==3.15.1
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.19.1
fonttools==4.25.0
fqdn==1.5.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
ipykernel==6.29.2
ipython==8.20.0
ipywidgets==8.1.2
isoduration==20.11.0
django-environ==0.11.2
jedi==0.18.1
Jinja2==3.1.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
kiwisolver==1.4.4
MarkupSafe==2.1.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.2
munkres==1.1.4
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.1.0
notebook_shim==0.2.4
numpy==1.26.3
openpyxl==3.1.2
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
pillow==10.2.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
psycopg2-binary==2.9.9
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
Send2Trash==1.8.2
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.2.0
terminado==0.18.0
typing==3.7.4.3
tinycss2==1.2.1
tornado==6.4
traitlets==5.7.1
types-python-dateutil==2.8.19.20240106
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.0
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
wheel==0.41.2
widgetsnbextension==4.0.10
faker==25.4.0
tqdm==4.66.4

# pytest
pluggy==1.5.0
iniconfig==2.0.0
pytest==8.2.2
pytest-django==4.8.0